cli: massive code refactor
Modularize the code to get rid of the massive types module, hopefully
making the code much easier to read.

Also removes dead code and crates that were not being used.
