+++ nix/cli/packages/cli/src/bitte/node/order.rs
@@ -1,0 +1,22 @@
+use super::BitteNode;
+use std::cmp::Ordering;
+
+impl Ord for BitteNode {
+    fn cmp(&self, other: &Self) -> Ordering {
+        self.name.cmp(&other.name)
+    }
+}
+
+impl PartialOrd for BitteNode {
+    fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
+        Some(self.cmp(other))
+    }
+}
+
+impl PartialEq for BitteNode {
+    fn eq(&self, other: &Self) -> bool {
+        self.name == other.name
+    }
+}
+
+impl Eq for BitteNode {}
