+++ nix/cli/packages/cli/src/nomad/mod.rs
@@ -1,0 +1,2 @@
+pub mod alloc;
+pub mod client;
